Fixed a lot of warnings to make the build cleaner, mostly type casts and vector iterators.

+++ Robot/controller/TrackingObject.h
@@ -41,9 +41,9 @@ public:
 		trackWindow_ = trackWindow;
 	}
 
-	float GetCenteredPercentage(int width);
+	int GetCenteredPercentage(int width);
 	float GetQuality(int width);
-	float GetSizePercentage();
+	int GetSizePercentage();
 
 	int CenterDistanceToDegrees(int width, bool cam);
